Name Redisson manager instrumentation by type
