+++ instrumentation/redisson/redisson-3.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/redisson/v3_0/RedissonInstrumentationModule.java
@@ -33,7 +33,7 @@ public class RedissonInstrumentationModule extends InstrumentationModule {
   public List<TypeInstrumentation> typeInstrumentations() {
     return asList(
         new ConnectionManagerConnectionFutureInstrumentation(),
-        new ConnectionManagerInstrumentation(),
+        new MasterSlaveConnectionManagerInstrumentation(),
         new RedisConnectionInstrumentation(),
         new RedisCommandDataInstrumentation());
   }
